// emmaus_ui/src/app/services/models/inquiry-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface InquiryResponse {
  createdAt?: string;
  message?: string;
  publicId?: string;
  remark?: string;
  senderEmail?: string;
  senderName?: string;
  senderPhone?: string;
  status?: string;
  subject?: string;
  type?: string;
  updatedAt?: string;
}